Rename filesio.py to fileio.py
